+++ public/init-redocly-ui.js
@@ -1,7 +1,7 @@
 'use strict';
 
-window.onload = () => {
+function loadRedocly(userOptions = {}) {
     const data = JSON.parse(document.getElementById('swagger-data').innerText);
 
-    Redoc.init(data.spec, {}, document.getElementById('swagger-ui'));
-};
+    Redoc.init(data.spec, userOptions, document.getElementById('swagger-ui'));
+}