+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 CHANGELOG
 =========
 
+4.26.0
+-----
+* Add ability to configure UI through configuration
+```yaml
+nelmio_api_doc:
+  html_config:
+    assets_mode: bundle
+    redocly_config:
+      expandResponses: '200,201'
+      hideDownloadButton: true
+    swagger_ui_config:
+      deepLinking: true
+```
+
 4.25.0
 -----
 * Added support for [JMS @Discriminator](https://jmsyst.com/libs/serializer/master/reference/annotations#discriminator) annotation/attribute